mito-ai: codebase refactor pt 2
