+++ mito-ai/src/Extensions/VariableManager/VariableInspector.tsx
